backend/requirements.txt
#file:ariadne/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.0.3
alembic==1.13.0
bcrypt==4.0.1
email-validator==2.1.0
cryptography==41.0.7
fastapi.middleware.cors